Support i18n:
I have added to home.html a title "My Home Page" and a button "Add Payments". The labels are taken via i18n and I needed to edit the settings to include additional middleware and LOCALE_PATHS
The message compile file is in the git .ignore so I didn't add and it should be created locally.